+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## pre-release v0.5
+
+- Removed **xISHServer** module from this repository and moved to [ISHServer](https://github.com/Sarafian/ISHServer).
+- Fixed some issues with the `Restart-Server.ps1` and `Test-Server.ps1` scripts.
+
 ## pre-release v0.4
 
 - All source code has header based on SDL's open source policy.